move to investor view
