+++ ui/src/components/Exchange/ExchangeParticipants.scss
@@ -1,0 +1,17 @@
+@import './CreateMarket.scss';
+
+@mixin exchange-participants {
+    .exchange-participants {
+        height: 100%;
+
+        a {
+            display: flex;
+            flex-direction: row;
+            align-items: center;
+
+            .icon {
+                margin-right: $spacing-xs;
+            }
+        }
+    }
+}
